separate RA protocols
